Starting to refactor into library
Migrating the core logic to a library file, still need to break the
logic out into modules.
